backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
sqlparse==0.5.3
dj-database-url==2.2.0
psycopg2-binary>=2.9
pytest
pytest-django
python-dotenv
openai>=1.0.0
requests
Pillow>=10.0.0
